Improves header navigation and mobile menu
Enhances header navigation by implementing smooth scrolling to sections, even after redirecting from other pages. It leverages sessionStorage to persist the target section and scrolls to it after the page loads.

The mobile menu is enhanced with an overlay to prevent interaction with the page behind it, and it now closes on any link click. Scrolling is also prevented on the body when the mobile menu is open for better UX.